third_party/blink/web_tests/external/wpt/editing/run/forwarddelete_6001-7000-expected.txt

This is a testharness.js-based test.
Found 25 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["forwarddelete",""]] "<ol><li>foo[<li>]bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foobar</li></ol>" but got "<ol><li>foo</li><li>bar</li></ol>"
[FAIL] [["forwarddelete",""]] "<ol><li>foo</ol><p>{}<br></p><ol><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>bar</p>" but got "<ol><li>foo</li></ol><ol><li>bar</li></ol>"
[FAIL] [["forwarddelete",""]] "<ol><li><p>foo</ol><p>{}<br></p><ol><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><p>foo</p></li></ol><p>bar</p>" but got "<ol><li><p>foo</p></li></ol><ol><li>bar</li></ol>"
[FAIL] [["forwarddelete",""]] "<ol><ol><li>foo</ol><li>{}<br><ol><li>bar</ol></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol><li>foo</li></ol><li>bar</li></ol>" but got "<ol><ol><li>foo</li></ol><li><ol><li>bar</li></ol></li></ol>"
[FAIL] [["forwarddelete",""]] "<ol><ol><li>foo</ol><li>{}<br></li><ol><li>bar</ol></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol><li>foo</li></ol><li>bar</li></ol>" but got "<ol><ol><li>foo</li></ol><ol><li>bar</li></ol></ol>"
[FAIL] [["forwarddelete",""]] "<ol><li>foo[</ol>bar]<ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>baz</li></ol>" but got "<ol><li>foo</li></ol><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["forwarddelete",""]] "<ol><li>foo[</ol><p>bar]<ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>baz</li></ol>" but got "<ol><li>foo</li></ol><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["forwarddelete",""]] "<ol><li>foo[</ol><p>bar]<ol><li>baz</ol>" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["defaultparagraphseparator","p"],["forwarddelete",""]] "<ol><li>foo[</ol><p>bar]<ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>baz</li></ol>" but got "<ol><li>foo</li></ol><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["forwarddelete",""]] "<ol><li><p>foo[</ol><p>bar]<ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><p>foo</p></li><li>baz</li></ol>" but got "<ol><li><p>foo</p></li></ol><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["forwarddelete",""]] "<ol><li><p>foo[</ol><p>bar]<ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><p>foo</p></li><li>baz</li></ol>" but got "<ol><li><p>foo</p></li></ol><ol><li>baz</li></ol>"
[FAIL] [["forwarddelete",""]] "<ol><li>foo</ol><p>[bar<ol><li>]baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>baz</p>" but got "<ol><li>foo</li></ol><ol><li>baz</li></ol>"
[FAIL] [["forwarddelete",""]] "<ol><ol><li>foo[</ol><li>bar</ol>baz]<ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol><li>foo</li></ol><li>quz</li></ol>" but got "<ol><ol><li>foo</li></ol></ol><ol><li>quz</li></ol>"
[FAIL] [["forwarddelete",""]] "<ul><li>foo</ul><p>{}<br></p><ul><li>bar</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><p>bar</p>" but got "<ul><li>foo</li></ul><ul><li>bar</li></ul>"
[FAIL] [["forwarddelete",""]] "<ol><li>foo[<li>bar]</ol><ol><li>baz</ol><ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>baz</li><li>quz</li></ol>" but got "<ol><li>foo</li></ol><ol><li>baz</li></ol><ol><li>quz</li></ol>"
[FAIL] [["forwarddelete",""]] "<ol><li>foo</ol><p>{}<br></p><ul><li>bar</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>bar</p>" but got "<ol><li>foo</li></ol><ul><li>bar</li></ul>"
[FAIL] [["forwarddelete",""]] "<ul><li>foo</ul><p>{}<br></p><ol><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><p>bar</p>" but got "<ul><li>foo</li></ul><ol><li>bar</li></ol>"
[FAIL] [["forwarddelete",""]] "<div><b>[foo]</b></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><br></div>" but got "<br>"
[FAIL] [["forwarddelete",""]] "foo[]<img contenteditable=false src=/img/lion.svg>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobar" but got "foo<img contenteditable=\\"false\\" src=\\"/img/lion.svg\\">bar"
[FAIL] [["forwarddelete",""]] "<div><div>{}<br></div></div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div><div><br></div></div>" not in array ["", "<br>"]
[FAIL] [["forwarddelete",""]] "<div><div contenteditable=false><div contenteditable><div>{}<br></div></div></div></div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div><div contenteditable=\\"false\\"><div contenteditable=\\"\\"><div><br></div></div></div></div>" not in array ["<div><div contenteditable=\\"false\\"><div contenteditable=\\"\\"></div></div></div>", "<div><div contenteditable=\\"false\\"><div contenteditable=\\"\\"><br></div></div></div>"]
[FAIL] [["forwarddelete",""]] "<div style=white-space:nowrap>[]f\\nbar</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"white-space:nowrap\\">bar</div>" but got "<div style=\\"white-space:nowrap\\">&nbsp;bar</div>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p><font color=blue>foo[]</font><p><font color=brown>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"blue\\">foo</font><font color=\\"brown\\">bar</font></p>" but got "<p><font color=\\"blue\\">foo</font><span style=\\"color:rgb(165, 42, 42)\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p><font color=blue>foo[]</font><p><font color=brown>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"blue\\">foo</font><font color=\\"brown\\">bar</font></p>" but got "<p><font color=\\"blue\\">foo</font><span style=\\"color:rgb(165, 42, 42)\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p><font size=3>foo[]</font><p><font size=5>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"3\\">foo</font><font size=\\"5\\">bar</font></p>" but got "<p><font size=\\"3\\">foo</font><span style=\\"font-size:x-large\\">bar</span></p>"
Harness: the test ran to completion.
